replace pendingRequests and olderShifts arrays with simple links
